+++ lib/main.ts
@@ -3,13 +3,11 @@
 import RootProvider from './root-provider';
 import providerPaths from './provider-list';
 
-module.exports = {
-  activate() {
-    this.providers = providerPaths.map(path => require(path)["default"])
-      .map(ProviderClass => new ProviderClass());
-  },
+export function activate() {
+  this.providers = providerPaths.map(path => require(path)["default"])
+    .map(ProviderClass => new ProviderClass());
+}
 
-  provide() {
-    return new RootProvider(this.providers);
-  }
+export function provide() {
+  return new RootProvider(this.providers);
 }
